kdePackages.extra-cmake-modules: propagate qdoc by default
A lot of things hard-require the ECM bits that use it now.

Also, disable strictDeps, because it somehow explodes unrelated
package checks in the sheer presence of qttools in nativeBuildInputs.

I don't know if this is a good solution, but cmake is also still
painfully incapable of dealing with cross correctly.